+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Editor config file
 - Advanced example
 - Event delays
+- Event throttling
 
 ### Changed
 - Allow multiple instances of image loader to be active